backend/requirements.txt
# Django & API
django
djangorestframework

# AI & Image
torch
diffusers
transformers
safetensors
sentencepiece
Pillow

# Environment
python-dotenv